requirements.txt
pyside6
numpy
pytesseract
opencv-python
python-Levenshtein
mss